renames hooks and moves to utils folder, updates to accept the calculator rather than it's id
