+++ src/components/Output.tsx
@@ -1,58 +1,26 @@
-import { getEffectiveRate, getMonthlyPayment, getRateByFrequency } from "../utils/calculators"
 import { PaymentDetails } from "../definitions/OutputTypes"
-import { PaymentSchedules, STATUS } from "../definitions/StringTypes"
+import { STATUS } from "../definitions/StringTypes"
 import MortgageSchedule from "./MortgageSchedule.tsx"
-import { paymentScheduleFrequencyMap } from "../utils/helpers"
 import OutputSummary from "./OutputSummary"
 
 import '../styles/output.css'
-import useOutputValuesFromCalculator from "../hooks/useOutputValuesFromCalculator.ts";
+import getOutputValuesFromCalculator from "../utils/getOutputValuesFromCalculator.ts";
+import getPaymentDetailsFromCalculator from "../utils/getPaymentDetailsFromCalculator.ts";
+import {CalculatorInputs} from "../definitions/CalculatorDefinitions.ts";
 
 interface OutputProps {
-  rate: number,
-  principal: number,
-  amortization: number,
-  paymentType: PaymentSchedules,
-  term: number,
-  calculatorId: number
+  calculator: CalculatorInputs
 }
 
 
 
 export default function Output({
-  rate,
-  principal,
-  amortization,
-  paymentType,
-  term,
-  calculatorId
+  calculator
 }: OutputProps) {
 
-  const output = useOutputValuesFromCalculator(calculatorId)
+  const output = getOutputValuesFromCalculator(calculator)
 
-  let paymentDetails: PaymentDetails | undefined = undefined
-
-  const effectiveRate = getEffectiveRate(rate)
-  const monthlyRate = getRateByFrequency(effectiveRate, 12)
-
-  const paymentValues = {
-    principal,
-    rate: monthlyRate,
-    amortization: amortization * 12 // to get the months
-  }
-
-  if (amortization > 0 && rate > 0) {
-    let monthlyPayment = getMonthlyPayment(paymentValues)
-    let scheduleRate = getRateByFrequency(effectiveRate, paymentScheduleFrequencyMap[paymentType])
-
-    paymentDetails = {
-      principal,
-      schedule: paymentType,
-      scheduleRate,
-      monthlyPayment,
-      termLength: term
-    }
-  }
+  let paymentDetails: PaymentDetails | undefined = getPaymentDetailsFromCalculator(calculator)
   
   return (
     <div className="calculator-outputs">
